# hyper: simplify handlers