@joshka 2025-03-27
There's definitely a performance degradation in the results, and I haven't investigated why yet. The goal of the changes was to make the source a bit more maintainable, and then work on improving the perf aspects on top of that in ways where the code which impacts performance can be a bit more obvious and measured (with test variants). Because hyper is also the base for many of the rust frameworks, having numbers for each of the tests allows this to be used as the framework overhead in those tests (using the "versus" field in the configs", which would help show which parts of the perf are just pure hyper and which are from framework overhead (or in some cases improvements).